requirements.txt
# LangGraph ve LangChain
langgraph==0.2.14

# Veritabanı
sqlalchemy==2.0.28
pgvector==0.2.5
psycopg2-binary==2.9.9
alembic==1.13.1

# API ve Web Arayüzleri
fastapi==0.109.2
uvicorn==0.27.1
gradio==4.19.2
streamlit==1.31.0
httpx==0.26.0
pydantic==2.6.1

# Azure OpenAI
openai==1.12.0
azure-identity==1.15.0

# Yardımcı araçlar
python-dotenv==1.0.1
pyyaml==6.0.1
numpy==1.26.3